requirements.txt
asgiref==3.3.1
Django==3.1.7
entrypoints==0.3
flake8==3.9.0
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
django-lint
django-ipware
